# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "PetSource API"
  },
  "host": "petsourceapi.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "user",
      "description": "Operations about user"
    },
    {
      "name": "pet",
      "description": "Operations about pets"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created user ID",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/user/exists": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Check if user with email exists",
        "operationId": "checkIfExists",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "email of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User exist",
            "schema": {
              "type": "string"
            }
          },
          "404": {
            "description": "User does not exist",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user info",
        "operationId": "userGetInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User information",
            "schema": {
              "$ref": "#/definitions/UserView"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/users/{userId}/pets": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Return all users pets",
        "operationId": "userGetPets",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user for which to fetch pets",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of user owned pet objects",
            "schema": {
              "$ref": "#/definitions/ArrayOfPets"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/users/{userId}/wallet": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Return users wallet",
        "operationId": "userGetWallet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user for which to fetch wallet",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Users wallet object",
            "schema": {
              "$ref": "#/definitions/Wallet"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      },
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create users wallet",
        "operationId": "userCreateWallet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of user for which to fetch wallet",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Create user wallet with params",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created wallet address",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/pets": {
      "post": {
        "tags": [
          "pet"
        ],
        "summary": "Create pet",
        "operationId": "createPet",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created pet object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet creation TX",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/pets/{txId}/status": {
      "get" : {
        "tags": [
          "pet"
        ],
        "summary": "Get pet store transaction statuts",
        "operationId": "petGetTransactionStatus",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "txId",
            "in": "path",
            "description": "ID of Pet",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "ID of user logged in",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Tx status",
            "schema": {
              "type": "string"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    },
    "/pets/{petId}": {
      "get": {
        "tags": [
          "pet"
        ],
        "summary": "Get pet info",
        "operationId": "petGetInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "petId",
            "in": "path",
            "description": "ID of Pet",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "ID of user logged in",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Pet information",
            "schema": {
              "$ref": "#/definitions/PetView"
            }
          }
        },
        "security": [
          {
            "authentication": []
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "authentication": {
      "type": "apiKey",
      "name": "x-access-token",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "firstName",
        "lastName",
        "email",
        "phoneNumber",
        "countryOfResidence",
        "authPlatform"
      ],
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "countryOfResidence": {
          "type": "string"
        },
        "authPlatform": {
          "type": "string",
          "enum": ["FB", "GOOGLE", "CIVIC"]
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "UserLogin": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ],
      "xml": {
        "name": "UserLogin"
      }
    },
    "UserView": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "countryOfResidence": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Pet": {
      "type": "object",
      "required": [
        "userID",
        "password",
        "name",
        "breed",
        "chipId"
      ],
      "properties": {
        "userID": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "sex": {
          "type": "string"
        },
        "birthDate": {
          "type": "string",
          "example": "2018-09-07",
          "format": "date"
        },
        "species": {
          "type": "string"
        },
        "breed": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "chipNumber": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "Pet"
      }
    },
    "PetView": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "sex": {
          "type": "string"
        },
        "birthDate": {
          "type": "string",
          "example": "2018-09-07",
          "format": "date"
        },
        "species": {
          "type": "string"
        },
        "breed": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "chipNumber": {
          "type": "string"
        }
      },
      "xml": {
        "name": "PetView"
      }
    },
    "ArrayOfPets": {
      "type": "array",
      "items": {
        "type": "object"
      }
    },
    "Wallet": {
      "type": "object",
      "properties": {
        "ethAddress": {
          "type": "string"
        },
        "ethBalance": {
          "type": "number"
        },
        "ptsBalance": {
          "type": "number"
        }
      }
    }
  }
}